refactor: move File.hpp and File.cpp to File subdirectory
