+++ tests/CMakeLists.txt
@@ -35,8 +35,8 @@ target_sources(tests
         "${PROJECT_SOURCE_DIR}/src/Utilities/Utilities.hpp"
         "${PROJECT_SOURCE_DIR}/src/Utilities/Utilities.cpp"
 
-        "${PROJECT_SOURCE_DIR}/src/Terminal/File.hpp"
-        "${PROJECT_SOURCE_DIR}/src/Terminal/File.cpp"
+        "${PROJECT_SOURCE_DIR}/src/File/File.hpp"
+        "${PROJECT_SOURCE_DIR}/src/File/File.cpp"
 
         "${PROJECT_SOURCE_DIR}/src/ScreenBuffer/ScreenBuffer.hpp"
         "${PROJECT_SOURCE_DIR}/src/ScreenBuffer/ScreenBuffer.cpp"
